Fix mainpage drap,drop function mainpage image drag function by @Imjaae #4
